fix(plugins): extract shared PathHelpers module
Move PATHS_CONFIG, DEFAULT_PATHS, read_file to shared module
used by both JsonSchema and Embed classes.

Fix grammar: "in none of" → "in any of"
